requirements.txt
pyspark[sql]
urllib3
pyyaml